fixes sharequote, share props
